Fix issue #2
On failure, compare_exchange_weak loads the actual value of *this into
the expected parameter. In our case, it loads the value of head_handle
into c_handle. This means we don't have to load the value from
head_handle into c_handle again at the start of the loop.