Fix Issue #4
Fixes access after free when releasing memory.

There is no straightforward way to release memory while popping, as
there may be threads holding the reference basically indefinitely before
accessing it again.

This forces us to deffer the releasing of memory until the whole LFStack
is destroyed, making it someone else's problem if destructor runs
concurrently with something from the LFStack.

To achieve this, we create a pool of "free" or empty nodes, making use
of things we already implemented, like pushing and popping from a linked
list.

The pool creates a second stack, which is used to store these empty
nodes, and instead of allocating new node on every push into the
LFStack, we try to retrieve an already allocated node from the pool
before allocating new node. This makes everything work correctly, while
minimizing the amount of new code needed.

The implementations of push and pop were separated into push_to_handle
and try_pop methods, which implement the push and pop into handle passed
as an argument. This allows us to use these methods to push and pop both
from the head_handle and from the pool_handle.

Also implemented a destructor, so that the memory is cleared and
deallocated after the LFstack is destroyed.